Milkshake.Data

Description

This submodule contains all type definitions and Dhall counterparts.

Documentation

data Virtual

A virtual target is one that is not backed up by a file, but rather by something that could be a file. One example would be an entry in an sqlite database. This may be a future feature.

Constructors

Virtual

Fields

name :: Text

exists :: Text

content :: Text

data Target

A target is either a file, some virtual content, or a phony target.

Constructors

File Text

Generic Virtual

Phony Text

data Action

An Action is a node in our workflow.

Constructors

Action

Fields

target :: [Target]

list of targets generated by executing the script

dependency :: [Target]

list of dependencies

script :: Maybe Text

the script to execute

type Generator = [Target] -> [Target] -> Maybe Text

Function type for generating a script to convert a Rule into a specific Target.

data Rule

A Rule is a parametric Action. Given a list of targets and dependencies, the generator creates the corresponding script.

Constructors

Rule

Fields

name :: Text

a unique name for this rule

gen :: Generator

the generator function for the script

data Call

The Call is like a function call, where the Rule is the function and $sel:target:Action and dependecy are the arguments.

Constructors

Call

Fields

name :: Text

the name of the rule to trigger

target :: [Target]

the targets

dependency :: [Target]

the dependencies

data Stmt

The Milkshake script is an unordered list of statements. The Stmt type encodes statements in a Milkshake script.

Constructors

StmtAction Action

StmtRule Rule

StmtCall Call

StmtInclude FilePath

StmtMain [FilePath]

StmtWatch Watch

stmt :: Decoder Stmt

To decode a list of Milkshake statements from the Dhall configuration use this decoder.

>>> input (list stmt) "(entangled.dhall).milkshake"

readStmts :: MonadIO m => FilePath -> m [Stmt]

Read a list of statements from a script.

data Config

Transposed data record of a list of Stmt.

Constructors

Config

Fields

rules :: Map Text Generator

triggers :: [Call]

actions :: [Action]

includes :: [FilePath]

mainTarget :: [FilePath]

watches :: [Watch]

stmtsToConfig :: [Stmt] -> Config

Groups a list of Stmt into a Config record.

readConfig :: MonadIO m => FilePath -> m Config

Read a script directly to Config record.

data Watch

A Watch is used to keep targets up-to-date when source files change.

Constructors

Watch

Fields

paths :: [Text]

lists of paths to monitor

target :: Target

target to build on file event