Milkshake.Run

Description

Contains functions to execute the script using Shake.

Documentation

targetPath :: Target -> Maybe FilePath

Get FilePath from Target

isFileTarget :: Target -> Bool

Checks if target is a file

enter :: Action -> Rules ()

Create Rules from an Action.

runScript :: Text -> Action ()

Helper function that creates a `Shake.Action ()` from a scriptlet.

fromCall :: Config -> Call -> Either MilkshakeError Action

Given a Config and a Call, creates an Action.

immediateActions :: Config -> Either MilkshakeError [Action]

Looks for actions that are immediately runnable. These are the plain Action statements, as well as calls to rules. The calls are expanded into actions, and a list of all actions is returned.

loadIncludes :: (MonadThrow m, MonadIO m) => Config -> m Config

Recursively loads include statements, until no includes are left.